feat(tmux-notify): swap wezterm extensions for tmux status indicators
Replace wezterm-notify and wezterm-pane-control with a tmux-native
agent status indicator. The new tmux-notify extension is gated on
$TMUX and writes per-pane @pi_status user options via tmux set-option,
which a tmux config can aggregate into window-level status-bar tabs.

- Add extensions/tmux-notify (working/done lifecycle on $TMUX_PANE)
- Remove extensions/wezterm-notify and extensions/wezterm-pane-control
- Drop dead WezTerm fallback path in extensions/debug
- Update extension count (53 → 52) and refresh docs/nav
