+++ docs/src/content/docs/extensions/overview.mdx
@@ -31,7 +31,7 @@ these categories:
 
 | Category | Count | What's in it |
 |----------|-------|--------------|
-| **Tools** | 19 | bash, edit, read, write, cd, web_fetch, web_search, subagent, teams, tasks, agent-commands, ask-user-question, background-task, output-styles, theme-selector, MCP adapter, session-memory, wezterm-pane-control |
+| **Tools** | 18 | bash, edit, read, write, cd, web_fetch, web_search, subagent, teams, tasks, agent-commands, ask-user-question, background-task, output-styles, theme-selector, MCP adapter, session-memory |
 | **UI & Display** | 6 | custom footer, git status, tool display, minimal skill display, session-namer, random-spinner |
 | **Commands** | 7 | command-prompt, command-expansion, context-fork, file-reference, health, shell-interpolation, skill-commands |
 | **Utilities** | 6 | cheatsheet, context-usage, debug, init, show-system-prompt, read-tool-enhanced |
