+++ docs/src/content/docs/index.mdx
@@ -1,6 +1,6 @@
 ---
 title: tallow
-description: An extensible, multi-model coding agent for your terminal. 53 extensions, 34 themes, multi-agent teams, and lifecycle hooks. Compatible with Claude Code.
+description: An extensible, multi-model coding agent for your terminal. 52 extensions, 34 themes, multi-agent teams, and lifecycle hooks. Compatible with Claude Code.
 template: splash
 head:
   - tag: title
@@ -9,7 +9,7 @@ hero:
   image:
     file: ../../../../assets/logo.jpg
     alt: tallow logo
-  tagline: An extensible coding agent for your terminal. 53 extensions, 34 themes, 9 agent templates — install only what you need.
+  tagline: An extensible coding agent for your terminal. 52 extensions, 34 themes, 9 agent templates — install only what you need.
   actions:
     - text: Get Started
       link: /getting-started/installation/
@@ -62,7 +62,7 @@ Run Claude, GPT-4, Gemini, or any OpenAI-compatible model from the same interfac
 
 ### Real extension system
 
-Not plugins bolted on — a proper extension architecture. Tools, TUI widgets, lifecycle hooks, event bus, custom commands. Extensions can render rich terminal UI, intercept tool calls, and compose with each other. All 51 available features are extensions.
+Not plugins bolted on — a proper extension architecture. Tools, TUI widgets, lifecycle hooks, event bus, custom commands. Extensions can render rich terminal UI, intercept tool calls, and compose with each other. All 52 available features are extensions.
 
 </div>
 
@@ -134,7 +134,6 @@ Intercept any point in the agent lifecycle — before/after tool execution, on t
   <a class="extension-tag extension-tag-link" href="/extensions/custom-footer/">custom-footer</a>
   <a class="extension-tag extension-tag-link" href="/extensions/tool-display/">tool-display</a>
   <a class="extension-tag extension-tag-link" href="/extensions/web-fetch-tool/">web_fetch</a>
-  <a class="extension-tag extension-tag-link" href="/extensions/wezterm-pane-control/">wezterm-pane-control</a>
   <a class="extension-tag extension-tag-link" href="/extensions/cd-tool/">cd</a>
   <a class="extension-tag extension-tag-link" href="/extensions/init/">init</a>
   <a class="extension-tag extension-tag-link" href="/extensions/show-system-prompt/">show-system-prompt</a>
