--- extensions/wezterm-notify/extension.json
@@ -1,32 +1,0 @@
-{
-	"name": "wezterm-notify",
-	"version": "0.1.0",
-	"description": "Signal agent lifecycle status to WezTerm via OSC 1337 user variables — enables tab spinner and done-color indicators.",
-	"whenToUse": "Use when WezTerm should receive agent-state signals for tab indicators.",
-	"capabilities": {
-		"events": [
-			"agent_end",
-			"agent_start",
-			"before_agent_start",
-			"input",
-			"session_shutdown",
-			"session_start"
-		]
-	},
-	"permissionSurface": {
-		"filesystem": "none",
-		"shell": false,
-		"network": false,
-		"subprocess": false
-	},
-	"category": "integration",
-	"tags": ["terminal", "wezterm", "notifications"],
-	"files": ["index.ts"],
-	"relationships": [
-		{
-			"type": "enhances",
-			"extension": "wezterm-pane-control"
-		}
-	],
-	"npmDependencies": {}
-}
