--- extensions/wezterm-pane-control/__tests__/index.test.ts
@@ -1,464 +1,0 @@
-import { afterEach, describe, expect, it } from "bun:test";
-import type { ExtensionAPI } from "@mariozechner/pi-coding-agent";
-import weztermPaneControl, {
-	buildSplitCommandArgs,
-	buildWeztermPaneGuidance,
-	executeWeztermAction,
-	filterPanesToCurrentTab,
-	hasExplicitPaneRequest,
-	hasSensitiveOutputMention,
-	isPaneCreatingAction,
-	unescapeText,
-	type WeztermCliResult,
-	type WeztermPaneInfo,
-} from "../index.js";
-
-const ORIGINAL_WEZTERM_PANE = process.env.WEZTERM_PANE;
-
-/**
- * Restore WEZTERM_PANE after each test.
- * @returns Nothing
- */
-afterEach(() => {
-	if (ORIGINAL_WEZTERM_PANE === undefined) {
-		delete process.env.WEZTERM_PANE;
-		return;
-	}
-	process.env.WEZTERM_PANE = ORIGINAL_WEZTERM_PANE;
-});
-
-/**
- * Create a minimal mock ExtensionAPI for registration tests.
- * @returns Mock API instance and captured calls
- */
-function createMockPi(): {
-	readonly api: ExtensionAPI;
-	readonly registerToolCalls: unknown[];
-	readonly onCalls: Array<{ readonly event: string; readonly handler: unknown }>;
-} {
-	const registerToolCalls: unknown[] = [];
-	const onCalls: Array<{ readonly event: string; readonly handler: unknown }> = [];
-
-	return {
-		registerToolCalls,
-		onCalls,
-		api: {
-			registerTool(tool: unknown) {
-				registerToolCalls.push(tool);
-			},
-			on(event: string, handler: unknown) {
-				onCalls.push({ event, handler });
-			},
-		} as unknown as ExtensionAPI,
-	};
-}
-
-/**
- * Build a pane fixture with sensible defaults.
- * @param overrides - Field overrides for the pane fixture
- * @returns Pane fixture
- */
-function pane(overrides: Partial<WeztermPaneInfo>): WeztermPaneInfo {
-	return {
-		pane_id: 10,
-		tab_id: 1,
-		title: "zsh",
-		cwd: "file:///tmp/project",
-		is_active: false,
-		is_zoomed: false,
-		size: { rows: 40, cols: 120 },
-		...overrides,
-	};
-}
-
-/**
- * Create a CLI result object for mocked wezterm calls.
- * @param stdout - Command stdout
- * @param stderr - Command stderr
- * @param status - Exit status code
- * @returns Mock CLI result
- */
-function cli(stdout: string, stderr = "", status = 0): WeztermCliResult {
-	return { stdout, stderr, status };
-}
-
-describe("wezterm-pane-control registration", () => {
-	it("does not register when WEZTERM_PANE is missing", () => {
-		delete process.env.WEZTERM_PANE;
-		const mockPi = createMockPi();
-
-		weztermPaneControl(mockPi.api);
-
-		expect(mockPi.registerToolCalls.length).toBe(0);
-		expect(mockPi.onCalls.length).toBe(0);
-	});
-});
-
-describe("buildWeztermPaneGuidance", () => {
-	it("includes bg_bash-first default, pane exceptions, sending, and privacy guidance", () => {
-		const guidance = buildWeztermPaneGuidance(116);
-		expect(guidance).toContain("WezTerm pane 116");
-		// bg_bash-first default
-		expect(guidance).toContain("use bg_bash");
-		expect(guidance).toContain("bg_bash");
-		// Pane exceptions — exactly 2
-		expect(guidance).toContain("User explicitly requests it");
-		expect(guidance).toContain("Sensitive output");
-		// Anti-speculative clause
-		expect(guidance).toContain("Never open a pane speculatively");
-		// bg_bash handles everything
-		expect(guidance).toContain(
-			"bg_bash\nhandles dev servers, builds, watchers, TUI apps, progress bars"
-		);
-		// Sending commands
-		expect(guidance).toContain("appending \\n");
-		// Privacy
-		expect(guidance).toContain("Do NOT call read_text on that pane");
-		expect(guidance).toContain("LLM must not consume secrets");
-	});
-
-	it("lists exactly 2 numbered exceptions", () => {
-		const guidance = buildWeztermPaneGuidance(116);
-		const numberedLines = guidance.split("\n").filter((l) => /^\d+\.\s/.test(l));
-		expect(numberedLines.length).toBe(2);
-	});
-
-	it("does not contain removed over-permissive guidance", () => {
-		const guidance = buildWeztermPaneGuidance(116);
-		expect(guidance).not.toContain("Long-running process the user wants to visually monitor");
-		expect(guidance).not.toContain("Process that needs a proper shell environment");
-		expect(guidance).not.toContain("ask_user_question");
-		expect(guidance).not.toContain("just create the pane and run it");
-		// Removed TTY exception
-		expect(guidance).not.toContain("Interactive TTY required");
-		expect(guidance).not.toContain("curses/TUI interface");
-		expect(guidance).not.toContain("progress bars that require a real terminal");
-	});
-});
-
-describe("filterPanesToCurrentTab", () => {
-	it("returns only panes from the active tab", () => {
-		const panes = [
-			pane({ pane_id: 10, tab_id: 1 }),
-			pane({ pane_id: 11, tab_id: 1 }),
-			pane({ pane_id: 20, tab_id: 2 }),
-		];
-
-		const result = filterPanesToCurrentTab(panes, 10);
-
-		expect(result.tabId).toBe(1);
-		expect(result.panes.map((p) => p.pane_id)).toEqual([10, 11]);
-	});
-});
-
-describe("buildSplitCommandArgs", () => {
-	it("maps split params to wezterm cli args", () => {
-		const args = buildSplitCommandArgs(
-			{
-				direction: "right",
-				percent: 30,
-				command: ["tallow", "--model", "claude-sonnet-4-20250514"],
-				cwd: "/Users/kevin/dev/tallow",
-			},
-			18
-		);
-
-		expect(args).toEqual([
-			"split-pane",
-			"--pane-id",
-			"18",
-			"--right",
-			"--percent",
-			"30",
-			"--cwd",
-			"/Users/kevin/dev/tallow",
-			"--",
-			"tallow",
-			"--model",
-			"claude-sonnet-4-20250514",
-		]);
-	});
-});
-
-describe("executeWeztermAction", () => {
-	it("list filters panes to the current tab", () => {
-		const listJson = JSON.stringify([
-			pane({ pane_id: 10, tab_id: 1, title: "main" }),
-			pane({ pane_id: 11, tab_id: 1, title: "logs" }),
-			pane({ pane_id: 20, tab_id: 2, title: "other" }),
-		]);
-
-		const result = executeWeztermAction(
-			{ action: "list" },
-			{
-				currentPaneId: 10,
-				runCli: () => cli(listJson),
-			}
-		);
-
-		expect(result.isError).toBeUndefined();
-		expect(result.content[0].text).toContain("pane 10");
-		expect(result.content[0].text).toContain("pane 11");
-		expect(result.content[0].text).not.toContain("pane 20");
-		expect(result.details?.tabId).toBe(1);
-	});
-
-	it("split builds expected cli invocation", () => {
-		let capturedArgs: readonly string[] = [];
-		const result = executeWeztermAction(
-			{
-				action: "split",
-				direction: "right",
-				percent: 25,
-				command: ["tallow"],
-			},
-			{
-				currentPaneId: 18,
-				runCli: (args) => {
-					capturedArgs = args;
-					return cli("52\n");
-				},
-			}
-		);
-
-		expect(capturedArgs).toEqual([
-			"split-pane",
-			"--pane-id",
-			"18",
-			"--right",
-			"--percent",
-			"25",
-			"--",
-			"tallow",
-		]);
-		expect(result.content[0].text).toContain("new pane 52");
-	});
-
-	it("close defaults to current pane and includes self-close warning", () => {
-		let capturedArgs: readonly string[] = [];
-		const result = executeWeztermAction(
-			{ action: "close" },
-			{
-				currentPaneId: 18,
-				runCli: (args) => {
-					capturedArgs = args;
-					return cli("");
-				},
-			}
-		);
-
-		expect(capturedArgs).toEqual(["kill-pane", "--pane-id", "18"]);
-		expect(result.content[0].text).toContain("Closing current pane (pane 18)");
-		expect(result.content[0].text).toContain("✓ Closed pane 18");
-	});
-
-	it("close targets explicit pane id", () => {
-		let capturedArgs: readonly string[] = [];
-		const result = executeWeztermAction(
-			{ action: "close", paneId: 25 },
-			{
-				currentPaneId: 18,
-				runCli: (args) => {
-					capturedArgs = args;
-					return cli("");
-				},
-			}
-		);
-
-		expect(capturedArgs).toEqual(["kill-pane", "--pane-id", "25"]);
-		expect(result.content[0].text).toBe("✓ Closed pane 25");
-	});
-
-	it("focus by direction uses activate-pane-direction", () => {
-		let capturedArgs: readonly string[] = [];
-		const result = executeWeztermAction(
-			{ action: "focus", direction: "left" },
-			{
-				currentPaneId: 18,
-				runCli: (args) => {
-					capturedArgs = args;
-					return cli("");
-				},
-			}
-		);
-
-		expect(capturedArgs).toEqual(["activate-pane-direction", "--pane-id", "18", "Left"]);
-		expect(result.content[0].text).toContain("left");
-	});
-
-	it("focus by pane id uses activate-pane", () => {
-		let capturedArgs: readonly string[] = [];
-		const result = executeWeztermAction(
-			{ action: "focus", paneId: 42 },
-			{
-				currentPaneId: 18,
-				runCli: (args) => {
-					capturedArgs = args;
-					return cli("");
-				},
-			}
-		);
-
-		expect(capturedArgs).toEqual(["activate-pane", "--pane-id", "42"]);
-		expect(result.content[0].text).toContain("pane 42");
-	});
-
-	it("returns tool error when wezterm cli command fails", () => {
-		const result = executeWeztermAction(
-			{ action: "move_to_tab", paneId: 999 },
-			{
-				currentPaneId: 18,
-				runCli: () => cli("", "pane not found", 1),
-			}
-		);
-
-		expect(result.isError).toBe(true);
-		expect(result.content[0].text).toContain("pane not found");
-	});
-
-	it("send_text unescapes \\n and uses --no-paste", () => {
-		let capturedArgs: readonly string[] = [];
-		const result = executeWeztermAction(
-			{ action: "send_text", text: "pnpm dev\\n" },
-			{
-				currentPaneId: 18,
-				runCli: (args) => {
-					capturedArgs = args;
-					return cli("");
-				},
-			}
-		);
-
-		expect(result.isError).toBeUndefined();
-		expect(capturedArgs).toEqual(["send-text", "--no-paste", "--pane-id", "18", "pnpm dev\n"]);
-	});
-
-	it("send_text unescapes hex escapes like \\x03 (Ctrl-C)", () => {
-		let capturedArgs: readonly string[] = [];
-		executeWeztermAction(
-			{ action: "send_text", text: "\\x03" },
-			{
-				currentPaneId: 18,
-				runCli: (args) => {
-					capturedArgs = args;
-					return cli("");
-				},
-			}
-		);
-
-		expect(capturedArgs).toEqual(["send-text", "--no-paste", "--pane-id", "18", "\x03"]);
-	});
-
-	it("send_text preserves literal backslashes via \\\\", () => {
-		let capturedArgs: readonly string[] = [];
-		executeWeztermAction(
-			{ action: "send_text", text: "echo \\\\n" },
-			{
-				currentPaneId: 18,
-				runCli: (args) => {
-					capturedArgs = args;
-					return cli("");
-				},
-			}
-		);
-
-		// \\n → literal backslash followed by 'n' (not a newline)
-		expect(capturedArgs[4]).toBe("echo \\n");
-	});
-});
-
-describe("unescapeText", () => {
-	it("converts \\n to newline", () => {
-		expect(unescapeText("hello\\n")).toBe("hello\n");
-	});
-
-	it("converts \\t to tab", () => {
-		expect(unescapeText("col1\\tcol2")).toBe("col1\tcol2");
-	});
-
-	it("converts \\r to carriage return", () => {
-		expect(unescapeText("line\\r")).toBe("line\r");
-	});
-
-	it("converts \\xNN hex escapes", () => {
-		expect(unescapeText("\\x03")).toBe("\x03");
-		expect(unescapeText("\\x1b")).toBe("\x1b");
-		expect(unescapeText("\\x41")).toBe("A");
-	});
-
-	it("converts \\\\\\\\ to literal backslash", () => {
-		expect(unescapeText("\\\\")).toBe("\\");
-	});
-
-	it("preserves \\\\ before escape chars: \\\\n → literal backslash + n", () => {
-		expect(unescapeText("\\\\n")).toBe("\\n");
-	});
-
-	it("leaves unrecognized sequences untouched", () => {
-		expect(unescapeText("hello\\z world")).toBe("hello\\z world");
-	});
-
-	it("handles multiple escapes in one string", () => {
-		expect(unescapeText("cd /tmp\\n\\nls -la\\n")).toBe("cd /tmp\n\nls -la\n");
-	});
-
-	it("returns empty string unchanged", () => {
-		expect(unescapeText("")).toBe("");
-	});
-
-	it("returns plain text unchanged", () => {
-		expect(unescapeText("no escapes here")).toBe("no escapes here");
-	});
-});
-
-describe("isPaneCreatingAction", () => {
-	it("returns true for pane-creating actions", () => {
-		expect(isPaneCreatingAction("split")).toBe(true);
-		expect(isPaneCreatingAction("spawn_tab")).toBe(true);
-		expect(isPaneCreatingAction("move_to_tab")).toBe(true);
-	});
-
-	it("returns false for non-creating actions", () => {
-		expect(isPaneCreatingAction("list")).toBe(false);
-		expect(isPaneCreatingAction("read_text")).toBe(false);
-		expect(isPaneCreatingAction("send_text")).toBe(false);
-		expect(isPaneCreatingAction("close")).toBe(false);
-		expect(isPaneCreatingAction("focus")).toBe(false);
-		expect(isPaneCreatingAction("zoom")).toBe(false);
-		expect(isPaneCreatingAction("resize")).toBe(false);
-	});
-});
-
-describe("hasExplicitPaneRequest", () => {
-	it("returns true for explicit pane/tab requests", () => {
-		expect(hasExplicitPaneRequest("open a pane for the server")).toBe(true);
-		expect(hasExplicitPaneRequest("new tab please")).toBe(true);
-		expect(hasExplicitPaneRequest("split the terminal")).toBe(true);
-		expect(hasExplicitPaneRequest("use wezterm to show logs")).toBe(true);
-		expect(hasExplicitPaneRequest("spawn a tab for this")).toBe(true);
-	});
-
-	it("returns false for non-pane prompts", () => {
-		expect(hasExplicitPaneRequest("start the dev server")).toBe(false);
-		expect(hasExplicitPaneRequest("run the build")).toBe(false);
-		expect(hasExplicitPaneRequest("left align the text")).toBe(false);
-		expect(hasExplicitPaneRequest("watch for file changes")).toBe(false);
-		expect(hasExplicitPaneRequest("open the browser")).toBe(false);
-	});
-});
-
-describe("hasSensitiveOutputMention", () => {
-	it("returns true for sensitive output mentions", () => {
-		expect(hasSensitiveOutputMention("show me the api key")).toBe(true);
-		expect(hasSensitiveOutputMention("generate a token")).toBe(true);
-		expect(hasSensitiveOutputMention("display the password")).toBe(true);
-		expect(hasSensitiveOutputMention("fetch my credentials")).toBe(true);
-		expect(hasSensitiveOutputMention("read the secret from 1password")).toBe(true);
-	});
-
-	it("returns false for non-sensitive prompts", () => {
-		expect(hasSensitiveOutputMention("start the dev server")).toBe(false);
-		expect(hasSensitiveOutputMention("run the tests")).toBe(false);
-		expect(hasSensitiveOutputMention("build the project")).toBe(false);
-	});
-});
