--- extensions/wezterm-pane-control/index.ts
@@ -1,820 +1,0 @@
-/**
- * WezTerm Pane Control Extension
- *
- * Adds a single `wezterm_pane` tool for prompt-driven pane management in
- * WezTerm. Registration is gated behind `WEZTERM_PANE` so non-WezTerm terminals
- * get zero overhead.
- */
-
-import { spawnSync } from "node:child_process";
-import { existsSync } from "node:fs";
-import { join } from "node:path";
-import { StringEnum, Type } from "@mariozechner/pi-ai";
-import type { ExtensionAPI } from "@mariozechner/pi-coding-agent";
-
-export type WeztermAction =
-	| "list"
-	| "split"
-	| "close"
-	| "focus"
-	| "zoom"
-	| "resize"
-	| "send_text"
-	| "read_text"
-	| "spawn_tab"
-	| "move_to_tab";
-
-export type WeztermDirection =
-	| "left"
-	| "right"
-	| "top"
-	| "bottom"
-	| "up"
-	| "down"
-	| "next"
-	| "prev";
-
-export type WeztermZoomState = "zoom" | "unzoom" | "toggle";
-
-export interface WeztermPaneParams {
-	action: WeztermAction;
-	paneId?: number;
-	direction?: WeztermDirection;
-	percent?: number;
-	command?: readonly string[];
-	cwd?: string;
-	text?: string;
-	state?: WeztermZoomState;
-	amount?: number;
-	startLine?: number;
-	endLine?: number;
-	escapes?: boolean;
-}
-
-export interface WeztermPaneInfo {
-	readonly pane_id: number;
-	readonly tab_id: number;
-	readonly title: string;
-	readonly cwd: string;
-	readonly is_active: boolean;
-	readonly is_zoomed: boolean;
-	readonly size: {
-		readonly rows: number;
-		readonly cols: number;
-	};
-}
-
-export interface WeztermCliResult {
-	readonly status: number;
-	readonly stdout: string;
-	readonly stderr: string;
-}
-
-export type WeztermCliRunner = (args: readonly string[]) => WeztermCliResult;
-
-interface ExecuteDeps {
-	readonly currentPaneId: number;
-	readonly runCli: WeztermCliRunner;
-}
-
-interface ToolResult {
-	content: Array<{ type: "text"; text: string }>;
-	details: Record<string, unknown>;
-	isError?: boolean;
-}
-
-/**
- * WezTerm pane actions that create new panes or tabs.
- * These are gated by the guardrail hook — blocked unless the user's prompt
- * explicitly requests pane/tab creation.
- */
-export const PANE_CREATING_ACTIONS: ReadonlySet<string> = new Set([
-	"split",
-	"spawn_tab",
-	"move_to_tab",
-]);
-
-/**
- * Patterns that indicate the user explicitly wants pane/tab creation.
- * Only unambiguous pane vocabulary — no generic words like "window", "left", "right".
- */
-export const EXPLICIT_PANE_REQUEST_PATTERNS: readonly RegExp[] = [
-	/\bwezterm\b/i,
-	/\bpane(?:s)?\b/i,
-	/\btab(?:s)?\b/i,
-	/\bsplit\b/i,
-	/\bspawn\b/i,
-];
-
-/**
- * Patterns that indicate sensitive output the LLM should not see.
- * Pane creation is allowed for these so the user can view secrets directly.
- */
-export const SENSITIVE_OUTPUT_PATTERNS: readonly RegExp[] = [
-	/\bsecret(?:s)?\b/i,
-	/\btoken(?:s)?\b/i,
-	/\bpassword(?:s)?\b/i,
-	/\bapi[- ]?key(?:s)?\b/i,
-	/\bcredential(?:s)?\b/i,
-];
-
-/**
- * Check whether a wezterm_pane action creates a new pane or tab.
- *
- * @param action - The action string from tool parameters
- * @returns True when the action would create a pane/tab
- */
-export function isPaneCreatingAction(action: string): boolean {
-	return PANE_CREATING_ACTIONS.has(action);
-}
-
-/**
- * Check whether user prompt text contains an explicit pane/tab request.
- *
- * @param prompt - The user's prompt text
- * @returns True when the prompt explicitly mentions panes/tabs
- */
-export function hasExplicitPaneRequest(prompt: string): boolean {
-	return EXPLICIT_PANE_REQUEST_PATTERNS.some((pattern) => pattern.test(prompt));
-}
-
-/**
- * Check whether user prompt text mentions sensitive output.
- *
- * @param prompt - The user's prompt text
- * @returns True when the prompt mentions secrets/tokens/passwords
- */
-export function hasSensitiveOutputMention(prompt: string): boolean {
-	return SENSITIVE_OUTPUT_PATTERNS.some((pattern) => pattern.test(prompt));
-}
-
-const ACTIONS: readonly WeztermAction[] = [
-	"list",
-	"split",
-	"close",
-	"focus",
-	"zoom",
-	"resize",
-	"send_text",
-	"read_text",
-	"spawn_tab",
-	"move_to_tab",
-] as const;
-
-const DIRECTIONS: readonly WeztermDirection[] = [
-	"left",
-	"right",
-	"top",
-	"bottom",
-	"up",
-	"down",
-	"next",
-	"prev",
-] as const;
-
-const ZOOM_STATES: readonly WeztermZoomState[] = ["zoom", "unzoom", "toggle"] as const;
-
-/**
- * Parse WEZTERM_PANE to a valid pane ID.
- *
- * @param value - Raw environment value
- * @returns Parsed pane ID, or null when missing/invalid
- */
-export function parsePaneId(value: string | undefined): number | null {
-	if (!value) return null;
-	const parsed = Number(value);
-	if (!Number.isInteger(parsed) || parsed <= 0) return null;
-	return parsed;
-}
-
-/**
- * Resolve the wezterm executable path.
- *
- * Honors WEZTERM_EXECUTABLE_DIR when available, otherwise falls back to
- * `wezterm` on PATH.
- *
- * @param env - Environment variables
- * @returns Executable path or binary name
- */
-export function resolveWeztermExecutable(env: NodeJS.ProcessEnv): string {
-	const executableDir = env.WEZTERM_EXECUTABLE_DIR;
-	if (typeof executableDir === "string" && executableDir.length > 0) {
-		const candidate = join(executableDir, "wezterm");
-		if (existsSync(candidate)) {
-			return candidate;
-		}
-	}
-	return "wezterm";
-}
-
-/**
- * Check whether wezterm CLI is available.
- *
- * @param executable - wezterm executable path/name
- * @returns True when wezterm CLI is invokable
- */
-export function isWeztermAvailable(executable: string): boolean {
-	const probe = spawnSync(executable, ["cli", "--help"], {
-		encoding: "utf-8",
-		stdio: ["ignore", "ignore", "ignore"],
-	});
-	if (probe.error) return false;
-	return probe.status === 0;
-}
-
-/**
- * Create a runner for `wezterm cli` commands.
- *
- * @param executable - wezterm executable path/name
- * @returns Runner function returning status/stdout/stderr
- */
-export function createWeztermRunner(executable: string): WeztermCliRunner {
-	return (args) => {
-		const result = spawnSync(executable, ["cli", ...args], {
-			encoding: "utf-8",
-			stdio: ["ignore", "pipe", "pipe"],
-		});
-
-		if (result.error) {
-			return {
-				status: 1,
-				stdout: "",
-				stderr: result.error.message,
-			};
-		}
-
-		return {
-			status: result.status ?? 1,
-			stdout: result.stdout ?? "",
-			stderr: result.stderr ?? "",
-		};
-	};
-}
-
-/**
- * Convert tool direction values to WezTerm's direction tokens.
- *
- * @param direction - User/tool direction string
- * @returns WezTerm direction token, or null when invalid
- */
-export function toPaneDirectionToken(
-	direction: string
-): "Left" | "Right" | "Up" | "Down" | "Next" | "Prev" | null {
-	switch (direction.toLowerCase()) {
-		case "left":
-			return "Left";
-		case "right":
-			return "Right";
-		case "top":
-		case "up":
-			return "Up";
-		case "bottom":
-		case "down":
-			return "Down";
-		case "next":
-			return "Next";
-		case "prev":
-			return "Prev";
-		default:
-			return null;
-	}
-}
-
-/**
- * Convert split direction to one of WezTerm split flags.
- *
- * @param direction - Optional split direction
- * @returns Split direction for CLI flags, or null when invalid
- */
-export function toSplitDirection(direction?: string): "left" | "right" | "top" | "bottom" | null {
-	if (!direction) return "bottom";
-	switch (direction.toLowerCase()) {
-		case "left":
-		case "right":
-		case "top":
-		case "bottom":
-			return direction.toLowerCase() as "left" | "right" | "top" | "bottom";
-		case "up":
-			return "top";
-		case "down":
-			return "bottom";
-		default:
-			return null;
-	}
-}
-
-/**
- * Build split-pane CLI args.
- *
- * @param params - Split action parameters
- * @param currentPaneId - Active pane ID
- * @returns CLI args for `wezterm cli split-pane`
- * @throws Error when direction/percent are invalid
- */
-export function buildSplitCommandArgs(
-	params: Pick<WeztermPaneParams, "paneId" | "direction" | "percent" | "command" | "cwd">,
-	currentPaneId: number
-): string[] {
-	const splitDirection = toSplitDirection(params.direction);
-	if (!splitDirection) {
-		throw new Error("split requires direction: left/right/top/bottom (up/down aliases supported)");
-	}
-
-	if (
-		params.percent !== undefined &&
-		(!Number.isFinite(params.percent) || params.percent < 1 || params.percent > 99)
-	) {
-		throw new Error("split percent must be between 1 and 99");
-	}
-
-	const args = [
-		"split-pane",
-		"--pane-id",
-		String(params.paneId ?? currentPaneId),
-		`--${splitDirection}`,
-	];
-
-	if (params.percent !== undefined) {
-		args.push("--percent", String(params.percent));
-	}
-	if (params.cwd) {
-		args.push("--cwd", params.cwd);
-	}
-	if (params.command && params.command.length > 0) {
-		args.push("--", ...params.command);
-	}
-
-	return args;
-}
-
-/**
- * Filter pane list to the current tab.
- *
- * @param panes - Full pane list from wezterm
- * @param currentPaneId - Current pane ID
- * @returns Filtered panes and resolved tab ID
- */
-export function filterPanesToCurrentTab(
-	panes: readonly WeztermPaneInfo[],
-	currentPaneId: number
-): { readonly tabId: number | null; readonly panes: readonly WeztermPaneInfo[] } {
-	const currentPane = panes.find((pane) => pane.pane_id === currentPaneId);
-	if (!currentPane) {
-		return { tabId: null, panes };
-	}
-	return {
-		tabId: currentPane.tab_id,
-		panes: panes.filter((pane) => pane.tab_id === currentPane.tab_id),
-	};
-}
-
-/**
- * Format pane records for readable tool output.
- *
- * @param panes - Pane list to format
- * @param currentPaneId - Current pane ID
- * @param tabId - Current tab ID (if known)
- * @returns Human-readable pane summary
- */
-export function formatPaneList(
-	panes: readonly WeztermPaneInfo[],
-	currentPaneId: number,
-	tabId: number | null
-): string {
-	if (panes.length === 0) {
-		return tabId === null ? "No panes found." : `No panes found in tab ${tabId}.`;
-	}
-
-	const header = tabId === null ? "Panes (current tab unknown):" : `Panes in tab ${tabId}:`;
-	const lines = panes.map((pane) => {
-		const tags = [
-			pane.pane_id === currentPaneId ? "this pane" : "",
-			pane.is_active ? "active" : "",
-			pane.is_zoomed ? "zoomed" : "",
-		].filter(Boolean);
-		const tagSuffix = tags.length > 0 ? ` (${tags.join(", ")})` : "";
-		const cwd = pane.cwd.replace(/^file:\/\//, "") || "(unknown cwd)";
-		const title = pane.title || "(no title)";
-		return [
-			`- pane ${pane.pane_id}${tagSuffix}`,
-			`  title: ${title}`,
-			`  cwd: ${cwd}`,
-			`  size: ${pane.size.cols}x${pane.size.rows}`,
-		].join("\n");
-	});
-
-	return `${header}\n${lines.join("\n")}`;
-}
-
-/**
- * Process common escape sequences in send_text payloads.
- *
- * LLM tool calls pass `\n` as literal backslash-n (two chars), not as an
- * actual newline. This converts common escape sequences to their real
- * character equivalents so `send_text` behaves as documented.
- *
- * Supported: `\n`, `\t`, `\r`, `\\`, `\xNN` (hex byte).
- *
- * @param text - Raw text from tool parameter
- * @returns Text with escape sequences resolved
- */
-export function unescapeText(text: string): string {
-	return text.replace(/\\(n|t|r|\\|x[0-9a-fA-F]{2})/g, (_match, seq: string) => {
-		switch (seq) {
-			case "n":
-				return "\n";
-			case "t":
-				return "\t";
-			case "r":
-				return "\r";
-			case "\\":
-				return "\\";
-			default:
-				if (seq.startsWith("x")) {
-					return String.fromCharCode(parseInt(seq.slice(1), 16));
-				}
-				return `\\${seq}`;
-		}
-	});
-}
-
-/**
- * Build a successful tool result.
- *
- * @param text - Primary output text
- * @param details - Optional structured details
- * @returns Tool result object
- */
-function ok(text: string, details: Record<string, unknown> = {}): ToolResult {
-	return {
-		content: [{ type: "text", text }],
-		details,
-	};
-}
-
-/**
- * Build an error tool result.
- *
- * @param text - Error message
- * @param details - Optional structured details
- * @returns Error tool result object
- */
-function fail(text: string, details: Record<string, unknown> = {}): ToolResult {
-	return {
-		content: [{ type: "text", text }],
-		details,
-		isError: true,
-	};
-}
-
-/**
- * Run wezterm CLI and throw on failure.
- *
- * @param runCli - wezterm CLI runner
- * @param args - CLI args
- * @returns Trimmed stdout
- * @throws Error when command exits non-zero
- */
-function runOrThrow(runCli: WeztermCliRunner, args: readonly string[]): string {
-	const result = runCli(args);
-	if (result.status !== 0) {
-		const message =
-			result.stderr.trim() || `wezterm cli ${args[0]} failed with exit code ${result.status}`;
-		throw new Error(message);
-	}
-	return result.stdout.trim();
-}
-
-/**
- * Build system-prompt guidance for WezTerm pane control behavior.
- *
- * Encodes a bg_bash-first policy: panes are the exception, not the default.
- * Long-running processes (dev servers, watchers, builds) should use bg_bash.
- * Panes are reserved for explicit user requests and sensitive output that the
- * LLM must not see.
- *
- * @param currentPaneId - Current WezTerm pane ID
- * @returns Guidance block appended to the system prompt
- */
-export function buildWeztermPaneGuidance(currentPaneId: number): string {
-	return [
-		"# WezTerm Pane Control",
-		"",
-		`You are running in WezTerm pane ${currentPaneId}.`,
-		"Use the wezterm_pane tool to manage panes: split, close, focus, zoom, resize, send/read text, or spawn new tabs.",
-		'Use action "list" to see panes in the current tab.',
-		"",
-		"## Default: use bg_bash, not panes",
-		"",
-		"For long-running processes (dev servers, watchers, builds, tests, compilers),",
-		"use `bg_bash` with `background: true`. Monitor output with `task_output`.",
-		"This is the default — do NOT open a pane unless one of the exceptions below applies.",
-		"",
-		"## When to open a pane (exceptions only)",
-		"",
-		"Open a pane or tab ONLY when one of these is true:",
-		"",
-		'1. **User explicitly requests it** — they said "open a pane", "use a pane",',
-		'   "split", "new tab", or similar. Without an explicit request, do NOT open a pane.',
-		"2. **Sensitive output** — secrets, tokens, passwords the LLM must not see.",
-		"   Spawn the pane, send the command, do NOT read_text the output.",
-		"",
-		"If uncertain, ALWAYS use bg_bash. Never open a pane speculatively — bg_bash",
-		"handles dev servers, builds, watchers, TUI apps, progress bars, and",
-		"interactive prompts fine.",
-		"",
-		"## Sending commands",
-		"",
-		"When you send a command to a pane, execute it by appending \\n (Enter).",
-		"Only leave a command unexecuted when the user explicitly asks to review it first.",
-		"Escape sequences (\\n, \\t, \\r, \\xNN) in send_text are processed automatically.",
-		"",
-		"## Secrets and sensitive output",
-		"",
-		"When a command generates or displays sensitive values (API keys, tokens, encryption keys, passwords):",
-		"- Spawn a pane and send the command so the USER can see the output.",
-		"- Do NOT call read_text on that pane — the LLM must not consume secrets.",
-		"- Tell the user their output is in the pane and let them handle it.",
-		"- Only read sensitive pane output if the user explicitly asks you to.",
-	].join("\n");
-}
-
-/**
- * Execute one wezterm pane action.
- *
- * @param params - Tool parameters
- * @param deps - Execution dependencies
- * @returns Tool result for the action
- */
-export function executeWeztermAction(params: WeztermPaneParams, deps: ExecuteDeps): ToolResult {
-	try {
-		switch (params.action) {
-			case "list": {
-				const raw = runOrThrow(deps.runCli, ["list", "--format", "json"]);
-				const parsed = JSON.parse(raw) as WeztermPaneInfo[];
-				const { tabId, panes } = filterPanesToCurrentTab(parsed, deps.currentPaneId);
-				return ok(formatPaneList(panes, deps.currentPaneId, tabId), {
-					tabId,
-					panes,
-				});
-			}
-
-			case "split": {
-				const args = buildSplitCommandArgs(params, deps.currentPaneId);
-				const newPaneId = runOrThrow(deps.runCli, args);
-				const direction = toSplitDirection(params.direction) ?? "bottom";
-				const percentLabel = params.percent === undefined ? "default size" : `${params.percent}%`;
-				return ok(
-					`✓ Split pane (${direction}, ${percentLabel})${newPaneId ? ` → new pane ${newPaneId}` : ""}`,
-					{
-						newPaneId: newPaneId || null,
-						direction,
-						percent: params.percent ?? null,
-					}
-				);
-			}
-
-			case "close": {
-				const targetPaneId = params.paneId ?? deps.currentPaneId;
-				const warning =
-					targetPaneId === deps.currentPaneId
-						? `Closing current pane (pane ${targetPaneId}). This will terminate this session.`
-						: null;
-				runOrThrow(deps.runCli, ["kill-pane", "--pane-id", String(targetPaneId)]);
-				return ok(
-					warning ? `${warning}\n✓ Closed pane ${targetPaneId}` : `✓ Closed pane ${targetPaneId}`,
-					{
-						paneId: targetPaneId,
-						isSelfClose: targetPaneId === deps.currentPaneId,
-					}
-				);
-			}
-
-			case "focus": {
-				if (params.paneId !== undefined) {
-					runOrThrow(deps.runCli, ["activate-pane", "--pane-id", String(params.paneId)]);
-					return ok(`✓ Focused pane ${params.paneId}`, { paneId: params.paneId });
-				}
-				if (!params.direction) {
-					return fail("focus requires either paneId or direction");
-				}
-				const direction = toPaneDirectionToken(params.direction);
-				if (!direction) {
-					return fail("focus direction must be one of: left/right/top/bottom/up/down/next/prev");
-				}
-				runOrThrow(deps.runCli, [
-					"activate-pane-direction",
-					"--pane-id",
-					String(deps.currentPaneId),
-					direction,
-				]);
-				return ok(`✓ Focused pane ${direction.toLowerCase()}`, { direction });
-			}
-
-			case "zoom": {
-				const state = params.state ?? "toggle";
-				const targetPaneId = params.paneId ?? deps.currentPaneId;
-				runOrThrow(deps.runCli, ["zoom-pane", "--pane-id", String(targetPaneId), `--${state}`]);
-				return ok(`✓ Applied zoom state '${state}' on pane ${targetPaneId}`, {
-					paneId: targetPaneId,
-					state,
-				});
-			}
-
-			case "resize": {
-				if (!params.direction) {
-					return fail("resize requires direction");
-				}
-				const direction = toPaneDirectionToken(params.direction);
-				if (!direction) {
-					return fail("resize direction must be one of: left/right/top/bottom/up/down/next/prev");
-				}
-				const amount = params.amount ?? 5;
-				if (!Number.isFinite(amount) || amount <= 0) {
-					return fail("resize amount must be a positive number");
-				}
-				const targetPaneId = params.paneId ?? deps.currentPaneId;
-				runOrThrow(deps.runCli, [
-					"adjust-pane-size",
-					"--pane-id",
-					String(targetPaneId),
-					"--amount",
-					String(amount),
-					direction,
-				]);
-				return ok(`✓ Resized pane ${targetPaneId} ${direction.toLowerCase()} by ${amount} cells`, {
-					paneId: targetPaneId,
-					direction,
-					amount,
-				});
-			}
-
-			case "send_text": {
-				if (params.text === undefined || params.text.length === 0) {
-					return fail("send_text requires non-empty text");
-				}
-				const targetPaneId = params.paneId ?? deps.currentPaneId;
-				const processedText = unescapeText(params.text);
-
-				// Use --no-paste so newlines are sent as Enter keypresses
-				// (bracketed paste would insert them as literal newlines in the
-				// shell's edit buffer without executing).
-				runOrThrow(deps.runCli, [
-					"send-text",
-					"--no-paste",
-					"--pane-id",
-					String(targetPaneId),
-					processedText,
-				]);
-				return ok(`✓ Sent text to pane ${targetPaneId}`, { paneId: targetPaneId });
-			}
-
-			case "read_text": {
-				const targetPaneId = params.paneId ?? deps.currentPaneId;
-				const args = ["get-text", "--pane-id", String(targetPaneId)];
-				if (params.startLine !== undefined) {
-					args.push("--start-line", String(params.startLine));
-				}
-				if (params.endLine !== undefined) {
-					args.push("--end-line", String(params.endLine));
-				}
-				if (params.escapes === true) {
-					args.push("--escapes");
-				}
-				const text = runOrThrow(deps.runCli, args);
-				return ok(text.length > 0 ? text : "(empty pane text)", {
-					paneId: targetPaneId,
-				});
-			}
-
-			case "spawn_tab": {
-				const args = ["spawn", "--pane-id", String(deps.currentPaneId)];
-				if (params.cwd) {
-					args.push("--cwd", params.cwd);
-				}
-				if (params.command && params.command.length > 0) {
-					args.push("--", ...params.command);
-				}
-				const newPaneId = runOrThrow(deps.runCli, args);
-				return ok(`✓ Spawned new tab${newPaneId ? ` → pane ${newPaneId}` : ""}`, {
-					newPaneId: newPaneId || null,
-				});
-			}
-
-			case "move_to_tab": {
-				const targetPaneId = params.paneId ?? deps.currentPaneId;
-				runOrThrow(deps.runCli, ["move-pane-to-new-tab", "--pane-id", String(targetPaneId)]);
-				return ok(`✓ Moved pane ${targetPaneId} to a new tab`, { paneId: targetPaneId });
-			}
-
-			default: {
-				return fail(`Unsupported action: ${(params as { action: string }).action}`);
-			}
-		}
-	} catch (error) {
-		const message = error instanceof Error ? error.message : String(error);
-		return fail(`✗ Failed: ${message}`);
-	}
-}
-
-/**
- * Register the wezterm pane control tool when running inside WezTerm.
- *
- * @param pi - Extension API instance
- * @returns Nothing
- */
-export default function weztermPaneControl(pi: ExtensionAPI): void {
-	const currentPaneId = parsePaneId(process.env.WEZTERM_PANE);
-	if (currentPaneId === null) {
-		return;
-	}
-
-	const weztermExecutable = resolveWeztermExecutable(process.env);
-	if (!isWeztermAvailable(weztermExecutable)) {
-		return;
-	}
-
-	const runCli = createWeztermRunner(weztermExecutable);
-
-	pi.registerTool({
-		name: "wezterm_pane",
-		label: "wezterm_pane",
-		description: [
-			"Manage WezTerm panes and tabs from prompts.",
-			"Use action='list' to inspect panes in the current tab, then split/focus/close/zoom/resize/send_text/read_text/spawn_tab/move_to_tab.",
-			"Use best judgment around secrets and self-monitored commands: avoid auto-running or reading sensitive output unless explicitly requested.",
-		].join(" "),
-		parameters: Type.Object({
-			action: StringEnum(ACTIONS, { description: "Pane control action" }),
-			paneId: Type.Optional(
-				Type.Number({ description: "Target pane ID. Defaults to current pane." })
-			),
-			direction: Type.Optional(
-				StringEnum(DIRECTIONS, {
-					description: "Direction for split/focus/resize",
-				})
-			),
-			percent: Type.Optional(
-				Type.Number({ description: "Split percentage (1-99) for action='split'" })
-			),
-			command: Type.Optional(
-				Type.Array(Type.String(), {
-					description: "Command and args for split/spawn_tab, e.g. ['tallow', '--model', '...']",
-				})
-			),
-			cwd: Type.Optional(Type.String({ description: "Working directory for split/spawn_tab" })),
-			text: Type.Optional(
-				Type.String({
-					description: "Text payload for action='send_text'. Include \\n to press Enter.",
-				})
-			),
-			state: Type.Optional(
-				StringEnum(ZOOM_STATES, {
-					description: "Zoom state for action='zoom' (default: toggle)",
-				})
-			),
-			amount: Type.Optional(Type.Number({ description: "Resize amount in cells (default: 5)" })),
-			startLine: Type.Optional(Type.Number({ description: "Start line for action='read_text'" })),
-			endLine: Type.Optional(Type.Number({ description: "End line for action='read_text'" })),
-			escapes: Type.Optional(
-				Type.Boolean({ description: "Include escape sequences for action='read_text'" })
-			),
-		}),
-		async execute(_toolCallId, params, _signal, _onUpdate) {
-			return executeWeztermAction(params as WeztermPaneParams, { currentPaneId, runCli });
-		},
-	});
-
-	pi.on("before_agent_start", async (event) => {
-		return {
-			systemPrompt: `${event.systemPrompt}\n\n${buildWeztermPaneGuidance(currentPaneId)}`,
-		};
-	});
-
-	// Track the current turn's user prompt for the guardrail hook.
-	// Updated via "input" event which fires before agent processing.
-	let currentTurnPrompt = "";
-
-	pi.on("input", async (event) => {
-		currentTurnPrompt = event.text ?? "";
-	});
-
-	// Guardrail: block pane-creating actions unless the user explicitly asked for panes
-	// or mentioned sensitive output. This is the enforcement mechanism — prompt guidance
-	// alone has failed repeatedly (see plan 186 history).
-	pi.on("tool_call", async (event) => {
-		if (event.toolName !== "wezterm_pane") return;
-
-		const params = event.input as { action?: string } | undefined;
-		if (!params?.action || !isPaneCreatingAction(params.action)) return;
-
-		if (hasExplicitPaneRequest(currentTurnPrompt)) return;
-		if (hasSensitiveOutputMention(currentTurnPrompt)) return;
-
-		return {
-			block: true,
-			reason:
-				"Pane creation blocked — no explicit pane/tab request detected in this turn. Use bg_bash for long-running processes. The user must say 'pane', 'tab', 'split', etc. to open a WezTerm pane.",
-		};
-	});
-}
